changes and fixes

+++ Hypernex.CCK.Editor/TempDir.cs
@@ -8,18 +8,20 @@ namespace Hypernex.CCK.Editor
     public class TempDir : IDisposable
     {
         private const string TEMP_FOLDER = "Hypernex.CCK.Temp";
-        private const string ASSETS_TEMP_FOLDER = "Assets/Hypernex.CCK.Temp";
+        private const string ASSETS_TEMP_FOLDER = "Assets/Hypernex.CCK/Temp";
 
         private static readonly Dictionary<string, TempDir> TempDirInstances = new Dictionary<string, TempDir>();
 
         public static byte[] GetByteFromString(string s) => Encoding.UTF8.GetBytes(s);
 
+        private bool inAssets;
         private string guid;
         private string path;
 
-        public TempDir()
+        public TempDir(bool inAssets = false)
         {
-            /*if (inAssets)
+            this.inAssets = inAssets;
+            if (inAssets)
             {
                 if (!Directory.Exists(ASSETS_TEMP_FOLDER))
                     Directory.CreateDirectory(ASSETS_TEMP_FOLDER);
@@ -40,8 +42,8 @@ namespace Hypernex.CCK.Editor
                 path = Path.Combine(TEMP_FOLDER, guid);
                 if (!Directory.Exists(path))
                     Directory.CreateDirectory(path);
-            }*/
-            if (!Directory.Exists(TEMP_FOLDER))
+            }
+            /*if (!Directory.Exists(TEMP_FOLDER))
                 Directory.CreateDirectory(TEMP_FOLDER);
             guid = Guid.NewGuid().ToString();
             while(TempDirInstances.ContainsKey(guid))
@@ -49,7 +51,7 @@ namespace Hypernex.CCK.Editor
             path = Path.Combine(TEMP_FOLDER, guid);
             if (!Directory.Exists(path))
                 Directory.CreateDirectory(path);
-            TempDirInstances.Add(guid, this);
+            TempDirInstances.Add(guid, this);*/
         }
 
         public string GetPath() => path;
@@ -71,12 +73,12 @@ namespace Hypernex.CCK.Editor
 
         public void Dispose()
         {
-            /*if (inAssets)
+            if (inAssets)
             {
                 string n = Path.Combine(ASSETS_TEMP_FOLDER, guid + ".meta");
                 if(File.Exists(n))
                     File.Delete(n);
-            }*/
+            }
             if(Directory.Exists(path))
                 Directory.Delete(path, true);
             TempDirInstances.Remove(guid);
